requirements.txt
pyspellchecker
# Built with Python 3.11.0
# Uses standard libraries: logging, exceptions
